+++ animation_nodes/nodes/gpencil/offset_gp_stroke.py
@@ -1,8 +1,8 @@
 import bpy
 from bpy.props import *
-from . c_utils import offsetFloatList
+from .. number.c_utils import offsetFloats
 from ... events import executionCodeChanged
-from .. vector.offset_vector import offsetVector3DList
+from .. vector.c_utils import offset3DVectors
 from ... base_types import AnimationNode, VectorizedSocket
 from ... data_structures import VirtualVector3DList, VirtualDoubleList
 
@@ -29,17 +29,17 @@ class OffsetGPStrokeNode(bpy.types.Node, AnimationNode):
         self.newInput("GPStroke", "Stroke", "stroke", dataIsModified = True)
         self.newInput("Falloff", "Falloff", "falloff")
         self.newInput(VectorizedSocket("Vector", "useVectorList",
-            ("Offset Location", "offsetLocations", dict(value = (0, 0, 1))),
-            ("Offset Locations", "offsetLocations")))
+            ("Location", "offsetLocations", dict(value = (0, 0, 1))),
+            ("Locations", "offsetLocations")))
         self.newInput(VectorizedSocket("Float", "useStrengthList",
-            ("Offset Strength", "offsetStrength", dict(value = 1)),
-            ("Offset Strengths", "offsetStrengths")))
+            ("Strength", "offsetStrength", dict(value = 1)),
+            ("Strengths", "offsetStrengths")))
         self.newInput(VectorizedSocket("Float", "usePressureList",
-            ("Offset Pressure", "offsetPressure", dict(value = 1)),
-            ("Offset Pressures", "offsetPressures")))
+            ("Pressure", "offsetPressure", dict(value = 1)),
+            ("Pressures", "offsetPressures")))
         self.newInput(VectorizedSocket("Float", "useUVRotationList",
-            ("Offset UV-Rotation", "offsetUVRotation", dict(value = 1)),
-            ("Offset UV-Rotations", "offsetUVRotations")))
+            ("UV-Rotation", "offsetUVRotation", dict(value = 1)),
+            ("UV-Rotations", "offsetUVRotations")))
 
         self.newOutput("GPStroke", "Stroke", "stroke")
 
@@ -67,22 +67,23 @@ class OffsetGPStrokeNode(bpy.types.Node, AnimationNode):
             return stroke
 
         falloffEvaluator = self.getFalloffEvaluator(falloff)
+        influences = falloffEvaluator.evaluateList(stroke.vertices)
 
         if self.useStrength:
             _offsetStrengths = VirtualDoubleList.create(offsetStrengths, 0)
-            offsetFloatList(stroke.vertices, stroke.strengths, _offsetStrengths, falloffEvaluator)
+            offsetFloats(stroke.strengths, _offsetStrengths, influences)
 
         if self.usePressure:
             _offsetPressures = VirtualDoubleList.create(offsetPressures, 0)
-            offsetFloatList(stroke.vertices, stroke.pressures, _offsetPressures, falloffEvaluator)
+            offsetFloats(stroke.pressures, _offsetPressures, influences)
 
         if self.useUVRotation:
             _offsetUVRotations = VirtualDoubleList.create(offsetUVRotations, 0)
-            offsetFloatList(stroke.vertices, stroke.uvRotations, _offsetUVRotations, falloffEvaluator)
+            offsetFloats(stroke.uvRotations, _offsetUVRotations, influences)
 
         if self.useLocation:
             _offsetLocations = VirtualVector3DList.create(offsetLocations, (0, 0, 0))
-            offsetVector3DList(stroke.vertices, _offsetLocations, falloffEvaluator)
+            offset3DVectors(stroke.vertices, _offsetLocations, influences)
 
         return stroke