+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added Spline From GP Stroke node.
 - Added Offset GP Stroke node.
 - Added Polygon Indices List to Edge Indices List link conversion.
+- Added Offset Spline Node.
 
 ### Fixed